undo renaming of FBGBar
