+++ src/components/App/Game/GameOver.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { GameMode } from './GameModePicker';
 import { IGameArgs } from './GameBoardWrapper';
 import { GamesList } from '../GamesList';
-import FreeBoardGamesBar from '../FBGBar';
+import FreeBoardGamesBar from '../FreeBoardGamesBar';
 import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
 import ReplayIcon from '@material-ui/icons/Replay';
