+++ src/components/App/MessagePageClass.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import Button from '@material-ui/core/Button';
 import HomeIcon from '@material-ui/icons/Home';
 import CircularProgress from '@material-ui/core/CircularProgress';
-import FreeBoardGamesBar from './FBGBar';
+import FreeBoardGamesBar from './FreeBoardGamesBar';
 import SvgError from './media/SvgError';
 import Typography from '@material-ui/core/Typography';
 import Link from 'next/link';
