+++ src/pages/index.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import FreeBoardGamesBar from 'components/App/FBGBar';
+import FreeBoardGamesBar from 'components/App/FreeBoardGamesBar';
 import Header from 'components/Header';
 import { GamesList } from 'components/App/GamesList';
 import SEO from 'components/SEO';
